Decouple toolAvailability from UI concerns via externalToolDefs
toolAvailability.ts had 6 imports including vscode, axios, and
ToolCategory (a settings view schema) — mixing runtime checking
with UI metadata ownership. This happened because we merged the
definitions into the availability module to create SSOT.

Fix: extract EXTERNAL_TOOL_DEFS into src/tools/externalToolDefs.ts.
This file is the SSOT — it co-locates identity, check functions,
and dashboard metadata. The two consumers read what they need:

  externalToolDefs.ts  (SSOT: defs + checks + UI metadata)
         ↑                       ↑
  toolAvailability.ts     toolDashboardData.ts
  (2 imports: registry    (reads UI metadata
   + defs. Runs checks,   + check results)
   manages cache)

toolAvailability.ts drops from 6 imports to 2 — no more vscode,
axios, ToolCategory, getZoteroPort, or checkToolInstalled.
